requirements.txt
Flask==3.1.0
python-docx==1.1.2
pypdf==5.2.0
